# api/openapi.yaml
openapi: 3.0.3
info:
  title: boxincloud API
  version: 0.1.0-dev
  description: |
    Bibliothèque de BD, comics et mangas sur stockage objet.

    Contrat de l'API boxincloud.

    **Ce fichier est la source de vérité.** Le serveur Go, le client TypeScript
    du web et le client Dart de l'application Flutter en sont générés
    (`make generate`). Toute évolution de l'API commence ici — jamais par le
    code, sinon les clients divergent.

    Un test de contrat (`internal/httpapi/contract_test.go`) valide chaque
    réponse du serveur réel contre ce schéma : une divergence entre le contrat
    et l'implémentation fait échouer la CI.

    ## Erreurs

    Les erreurs suivent la [RFC 9457](https://www.rfc-editor.org/rfc/rfc9457)
    (`application/problem+json`). Le champ `type` est stable : les clients
    peuvent s'appuyer dessus pour brancher leur comportement. Deux valeurs
    méritent un traitement particulier :

    - `.../token-expired` — le jeton d'accès a expiré, rafraîchir puis rejouer ;
    - `.../session-revoked` — la session est révoquée, redemander une connexion
      complète (une réutilisation de refresh token a été détectée).

    ## Pagination

    Pagination par curseur. Le client renvoie `nextCursor` tel quel dans le
    paramètre `cursor` ; il ne doit jamais le construire ni l'interpréter.
    L'absence de `nextCursor` signifie « dernière page ».

    ## Versionnement

    Le préfixe `/api/v1` change uniquement en cas de rupture. Les ajouts
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /libraries:
    post:
      operationId: createLibrary
      summary: Créer une bibliothèque sur un backend existant
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, backendId]
              properties:
                name: { type: string }
                backendId: { type: string, format: uuid }
                kind:
                  type: string
                  description: "Défaut : comics."
                rootPrefix:
                  type: string
                  description: |
                    Préfixe sous lequel la bibliothèque vit dans le backend.
                    Vide pour la racine.
      responses:
        "201":
          description: Bibliothèque créée
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LibraryDetail" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"

    get:
      operationId: listLibraries
      summary: Bibliothèques visibles
      description: |
        Seules les bibliothèques auxquelles le compte a accès sont retournées.
        Une bibliothèque sans restriction est visible de tous.
      tags: [catalog]
      responses:
        "200":
          description: Bibliothèques
          content:
            application/json:
              schema:
                type: object
                required: [libraries]
                properties:
                  libraries:
                    type: array
                    items: { $ref: "#/components/schemas/Library" }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /comics/{comicId}/folder:
    put:
      operationId: moveComic
      summary: Ranger un album dans un autre dossier
      description: |
        Le dossier d'un album n'est pas une propriété du catalogue : il découle
        de la clé de l'objet dans le backend. Le déplacer renomme donc l'objet,
        par copie côté serveur quand le backend le permet — les octets ne
        transitent jamais par boxincloud.

        Une destination déjà occupée par un fichier du même nom est refusée
        plutôt qu'écrasée.
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/ComicId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [folder]
              properties:
                folder:
                  type: string
                  description: |
                    Dossier de destination, relatif au préfixe de la
                    bibliothèque. Vide pour la racine. Les remontées « .. » sont
                    neutralisées.
      responses:
        "200":
          description: Dossier réel après déplacement
          content:
            application/json:
              schema:
                type: object
                required: [folderPath]
                properties:
                  folderPath: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
  /comics/{comicId}:
    delete:
      operationId: deleteComic
      summary: Retirer un album
      description: |
        Le fichier est CONSERVÉ par défaut. Retirer un album d'un catalogue se
        rattrape ; effacer un fichier non.

        Sans le fichier, la ligne est marquée comme exclue plutôt que détruite :
        la progression de lecture, les favoris et les notes y sont rattachés, et
        les effacer priverait d'historique quelqu'un qui remettrait le fichier
        en place. Le parcours de bibliothèque respecte cette exclusion et ne
        fait pas réapparaître l'album.

        Avec le fichier, l'objet est supprimé du backend puis la ligne effacée.
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/ComicId"
        - name: deleteFile
          in: query
          description: Supprimer aussi le fichier du stockage. Irréversible.
          schema: { type: boolean }
      responses:
        "204":
          description: Album retiré
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Le dossier, ou l'un de ses parents, est en lecture seule
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }


    patch:
      operationId: editComic
      summary: Corriger les métadonnées
      description: |
        Chaque champ envoyé est appliqué **et verrouillé** : une réindexation ne
        l'écrasera plus. C'est la contrepartie indispensable de l'automatisme —
        sans elle, corriger un titre serait vain.
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/ComicId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                number: { type: string }
                summary: { type: string }
                language: { type: string }
      responses:
        "200":
          description: Album mis à jour
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comic" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    get:
      operationId: getComic
      summary: Détail d'un album
      description: |
        Un album d'une bibliothèque inaccessible répond 404, pas 403 : un 403
        confirmerait son existence à quelqu'un qui n'y a pas droit.
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/ComicId"
      responses:
        "200":
          description: Album
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comic" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Lecture ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Jeton d'accès obtenu via `POST /auth/login`. Durée de vie courte
        (15 min par défaut) ; à renouveler avec le refresh token.
  parameters:
    ComicId:
      name: comicId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    Library:
      type: object
      required: [id, name, kind, comicCount]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind:
          type: string
          enum: [comic, manga, book, mixed]
        comicCount: { type: integer, format: int32 }
    ComicState:
      type: string
      enum: [pending, indexing, ready, hydrating, error]
      description: |
        Seuls les albums `ready` sont lisibles. `error` porte une explication
        dans l'interface d'administration.
    Comic:
      type: object
      required: [id, libraryId, title, format, pageCount, state, fileSize, createdAt, coverPath, fileName, folderPath]
      properties:
        id: { type: string, format: uuid }
        libraryId: { type: string, format: uuid }
        seriesId: { type: string, format: uuid }
        seriesName: { type: string }
        title: { type: string }
        number:
          type: string
          description: |
            Texte, car les numéros réels sont sales — « 3 », « 3.5 », « HS ».
        volume: { type: integer, format: int32 }
        summary: { type: string }
        format:
          type: string
          enum: [cbz, cbr, cb7, pdf, epub]
        pageCount: { type: integer, format: int32 }
        state: { $ref: "#/components/schemas/ComicState" }
        ageRating: { type: integer, format: int32 }
        language: { type: string, description: Code BCP-47. }
        fileSize: { type: integer, format: int64 }
        releasedAt: { type: string, format: date }
        createdAt: { type: string, format: date-time }
        coverPath:
          type: string
          description: |
            Chemin de la couverture. Le client y ajoute `?width=` plutôt que de
            composer l'URL lui-même.
        coverPlaceholder:
          type: string
          description: |
            Aperçu de chargement (LQIP) : data-URI d'une image de 16 px de
            large, à afficher floutée le temps que la couverture arrive.
            Quelques centaines d'octets, livrés avec la liste pour éviter une
            requête supplémentaire.

            Absent tant que l'album n'a pas été indexé par une version qui le
            produit — le client doit gérer son absence.
        fileName:
          type: string
          description: Nom du fichier dans le backend, sans son chemin.
        folderPath:
          type: string
          description: |
            Dossier contenant l'album, relatif au préfixe de la bibliothèque.
            Vide à la racine.
...
    LibraryDetail:
      type: object
      required: [id, backendId, name, kind, rootPrefix, comicCount]
      properties:
        id: { type: string, format: uuid }
        backendId: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string }
        rootPrefix: { type: string }
        comicCount: { type: integer, format: int32 }
...
    Problem:
      type: object
      description: |
        Erreur au format RFC 9457, servie avec le type MIME
        `application/problem+json`.
      required: [type, title, status]
      properties:
        type:
          type: string
          description: Identifiant stable de la catégorie d'erreur.
        title: { type: string }
        status: { type: integer }
        detail:
          type: string
          description: Explication de ce cas précis, affichable à l'utilisateur.
        instance:
          type: string
          description: |
            Identifiant de la requête. À inclure dans un rapport de bug pour
            retrouver la trace côté serveur.
        errors:
          type: object
          additionalProperties: { type: string }
          description: Champs invalides, pour les erreurs de validation.
  responses:
...
    Unauthorized:
      description: Jeton absent, expiré ou invalide
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: Droits insuffisants
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Ressource introuvable
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationFailed:
      description: Validation échouée
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...